# handywedge-calendar/handywedge-calendar-core/src/main/java/com/handywedge/calendar/Office365/rest/models/EmailAddress.java
//package com.handywedge.calendar.rest.models;
//
//import java.util.Objects;
//
///**
// * EmailAddress
// */
//public class EmailAddress {
//
//  @Override
//  public boolean equals(java.lang.Object o) {
//    if (this == o) {
//      return true;
//    }
//    if (o == null || getClass() != o.getClass()) {
//      return false;
//    }
//    return true;
//  }
//
//  @Override
//  public int hashCode() {
//    return Objects.hash();
//  }
//
//
//  @Override
//  public String toString() {
//    StringBuilder sb = new StringBuilder();
//    sb.append("class EmailAddress {\n");
//    sb.append("}");
//    return sb.toString();
//  }
//
//  /**
//   * Convert the given object to string with each line indented by 4 spaces
//   * (except the first line).
//   */
//  private String toIndentedString(java.lang.Object o) {
//    if (o == null) {
//      return "null";
//    }
//    return o.toString().replace("\n", "\n    ");
//  }
//}